[MOVE] Moved code to ol_sale module
