+++ ls_pdf_reports/reports/__init__.py
@@ -4,6 +4,5 @@ from . import packing_slip_from_so
 from . import picking_list
 from . import picking_list_from_so
 from . import rma_repair_summary
-from . import sale_order_quote
 from . import inventory_sheets
 from . import automation_test
