+++ ol_sale/models/__init__.py
@@ -5,3 +5,5 @@ from . import sale_order
 from . import res_partner
 from . import stock_picking
 from . import mrp_production
+# TODO: "sale.order.payment.method" model not found
+# from . import payment_method
