+++ ol_sale/models/sale_order_line.py
@@ -141,4 +141,12 @@ class SaleOrderLine(models.Model):
         self.with_context(from_unlink=True).update_crm_tag_sale_order()
         return super().unlink()
 
+
+    def get_line_serial_numbers(self):
+        """
+        Get a list of serial numbers for this sale line
+        """
+        self.ensure_one()
+
+        return self.move_ids.move_line_ids.mapped('lot_id.name')
     # END #########
